refactor: modularize app and add classifier
