+++ app/main.py
@@ -1,32 +1,43 @@
 import os
 import sqlite3
 import time
-import json
 from typing import List, Dict, Optional
-
-from markdown import markdown
 
 import logging
 from collections import deque
 from threading import Thread
 
 import pandas as pd
-import requests
 from fastapi import FastAPI, Form, Request, BackgroundTasks
 from fastapi.responses import HTMLResponse, RedirectResponse, JSONResponse
 from fastapi.staticfiles import StaticFiles
 from fastapi.templating import Jinja2Templates
 from jobspy import scrape_jobs
 
-DATABASE = os.environ.get("DATABASE", "jobs.db")
-OLLAMA_BASE_URL = os.environ.get("OLLAMA_BASE_URL")
-
-# Allow separate models for embeddings and rephrasing while keeping backwards
-# compatibility with the original `OLLAMA_MODEL` variable.
-_default_model = os.environ.get("OLLAMA_MODEL", "llama3")
-OLLAMA_EMBED_MODEL = os.environ.get("OLLAMA_EMBED_MODEL", _default_model)
-OLLAMA_REPHRASE_MODEL = os.environ.get("OLLAMA_REPHRASE_MODEL", _default_model)
-OLLAMA_ENABLED = bool(OLLAMA_BASE_URL)
+from .config import (
+    DATABASE,
+    OLLAMA_ENABLED,
+    BUILD_NUMBER,
+)
+from .db import (
+    init_db,
+    save_jobs,
+    get_random_job,
+    get_job,
+    list_jobs_by_feedback,
+    aggregate_job_stats,
+    increment_rating_count,
+    record_feedback,
+    cleanup_jobs,
+)
+from .ai import (
+    ensure_model_downloaded,
+    embed_text,
+    generate_summary,
+    render_markdown,
+    process_all_jobs,
+)
+from .model import train_model, predict_unrated
 
 app = FastAPI()
 app.mount("/static", StaticFiles(directory="app/static"), name="static")
@@ -64,103 +75,6 @@ Job posting:
 '''
 
 
-def ensure_model_downloaded() -> None:
-    if not OLLAMA_ENABLED:
-        return
-    for model in {OLLAMA_EMBED_MODEL, OLLAMA_REPHRASE_MODEL}:
-        try:
-            requests.post(f"{OLLAMA_BASE_URL}/api/pull", json={"name": model}, timeout=120)
-        except Exception as exc:
-            logger.info(f"Failed to pull model {model}: {exc}")
-
-
-def embed_text(text: str) -> List[float]:
-    if not OLLAMA_ENABLED:
-        return []
-    try:
-        r = requests.post(
-            f"{OLLAMA_BASE_URL}/api/embeddings",
-            json={
-                "model": OLLAMA_EMBED_MODEL,
-                "prompt": text,
-                "options": {"num_ctx": 4096},
-            },
-            timeout=120,
-        )
-        r.raise_for_status()
-        return r.json().get("embedding", [])
-    except Exception as exc:
-        logger.info(f"Embedding failed: {exc}")
-        return []
-
-
-def generate_summary(text: str) -> str:
-    if not OLLAMA_ENABLED:
-        return ""
-    prompt = REWRITE_PROMPT_TEMPLATE.format(description=text)
-    try:
-        r = requests.post(
-            f"{OLLAMA_BASE_URL}/api/generate",
-            json={
-                "model": OLLAMA_REPHRASE_MODEL,
-                "prompt": prompt,
-                "stream": False,
-                "options": {"num_ctx": 8192},
-            },
-            timeout=120,
-        )
-        r.raise_for_status()
-        return r.json().get("response", "")
-    except Exception as exc:
-        logger.info(f"Summary generation failed: {exc}")
-        return ""
-
-
-def render_markdown(text: str) -> str:
-    """Convert Markdown to HTML after stripping '---' delimiters."""
-    if not text:
-        return ""
-    lines = [ln for ln in text.splitlines() if ln.strip() != "---"]
-    cleaned = "\n".join(lines)
-    fixed = []
-    for ln in cleaned.splitlines():
-        if ln.lstrip().startswith("-") and fixed and fixed[-1].strip():
-            fixed.append("")
-        fixed.append(ln)
-    cleaned = "\n".join(fixed)
-    return markdown(cleaned)
-
-
-def process_all_jobs() -> None:
-    if not OLLAMA_ENABLED:
-        return
-    conn = sqlite3.connect(DATABASE)
-    cur = conn.cursor()
-    cur.execute("SELECT id, description FROM jobs")
-    rows = cur.fetchall()
-    for job_id, desc in rows:
-        if not desc:
-            continue
-        cur.execute("SELECT 1 FROM summaries WHERE job_id=?", (job_id,))
-        have_sum = cur.fetchone()
-        cur.execute("SELECT 1 FROM embeddings WHERE job_id=?", (job_id,))
-        have_emb = cur.fetchone()
-        if have_sum and have_emb:
-            continue
-        summary = generate_summary(desc) if not have_sum else None
-        embedding = embed_text(desc) if not have_emb else None
-        if summary is not None:
-            cur.execute(
-                "INSERT OR IGNORE INTO summaries(job_id, summary) VALUES(?, ?)",
-                (job_id, summary),
-            )
-        if embedding is not None:
-            cur.execute(
-                "INSERT OR IGNORE INTO embeddings(job_id, embedding) VALUES(?, ?)",
-                (job_id, json.dumps(embedding)),
-            )
-        conn.commit()
-    conn.close()
 
 
 def format_salary(min_amount: float, max_amount: float, currency: str) -> str:
@@ -229,225 +143,6 @@ def log_progress(message: str) -> None:
 
 
 
-def init_db() -> None:
-    conn = sqlite3.connect(DATABASE)
-    cur = conn.cursor()
-    cur.execute(
-        """
-        CREATE TABLE IF NOT EXISTS jobs(
-            id INTEGER PRIMARY KEY AUTOINCREMENT,
-            site TEXT,
-            title TEXT,
-            company TEXT,
-            location TEXT,
-            date_posted TEXT,
-            description TEXT,
-            interval TEXT,
-            min_amount REAL,
-            max_amount REAL,
-            currency TEXT,
-            job_url TEXT UNIQUE,
-            rating_count INTEGER DEFAULT 0
-        )
-        """
-    )
-    cur.execute(
-        """
-        CREATE TABLE IF NOT EXISTS feedback(
-            id INTEGER PRIMARY KEY AUTOINCREMENT,
-            job_id INTEGER,
-            liked INTEGER,
-            reason TEXT,
-            rated_at INTEGER
-        )
-        """
-    )
-    cur.execute(
-        """
-        CREATE TABLE IF NOT EXISTS embeddings(
-            job_id INTEGER PRIMARY KEY,
-            embedding TEXT
-        )
-        """
-    )
-    cur.execute(
-        """
-        CREATE TABLE IF NOT EXISTS summaries(
-            job_id INTEGER PRIMARY KEY,
-            summary TEXT
-        )
-        """
-    )
-    conn.commit()
-    conn.close()
-
-
-def save_jobs(df: pd.DataFrame) -> None:
-    if df.empty:
-        return
-    cols = {
-        "site": None,
-        "title": None,
-        "company": None,
-        "location": None,
-        "date_posted": None,
-        "description": None,
-        "interval": None,
-        "min_amount": None,
-        "max_amount": None,
-        "currency": None,
-        "job_url": None,
-    }
-    # Keep only the columns we know about
-    df = df.loc[:, df.columns.intersection(cols.keys())]
-    conn = sqlite3.connect(DATABASE)
-    cur = conn.cursor()
-    for _, row in df.iterrows():
-        values = tuple(row.get(c) for c in cols)
-        cur.execute(
-            """
-            INSERT OR IGNORE INTO jobs(site,title,company,location,date_posted,description,interval,min_amount,max_amount,currency,job_url)
-            VALUES(?,?,?,?,?,?,?,?,?,?,?)
-            """,
-            values,
-        )
-    conn.commit()
-    conn.close()
-    if OLLAMA_ENABLED:
-        process_all_jobs()
-
-
-def get_random_job() -> Optional[Dict]:
-    conn = sqlite3.connect(DATABASE)
-    cur = conn.cursor()
-    cur.execute(
-        """
-        SELECT j.*, s.summary FROM jobs j
-        LEFT JOIN summaries s ON j.id = s.job_id
-        ORDER BY rating_count ASC, RANDOM() LIMIT 1
-        """
-    )
-    row = cur.fetchone()
-    columns = [c[0] for c in cur.description]
-    conn.close()
-    if row:
-        job = dict(zip(columns, row))
-        if job.get("summary"):
-            job["summary"] = render_markdown(job["summary"])
-        return job
-    return None
-
-
-def get_job(job_id: int) -> Optional[Dict]:
-    """Return a job by id including any summary."""
-    conn = sqlite3.connect(DATABASE)
-    cur = conn.cursor()
-    cur.execute(
-        """
-        SELECT j.*, s.summary FROM jobs j
-        LEFT JOIN summaries s ON j.id = s.job_id
-        WHERE j.id=?
-        """,
-        (job_id,),
-    )
-    row = cur.fetchone()
-    columns = [c[0] for c in cur.description]
-    conn.close()
-    if row:
-        job = dict(zip(columns, row))
-        if job.get("summary"):
-            job["summary"] = render_markdown(job["summary"])
-        return job
-    return None
-
-
-def list_jobs_by_feedback() -> List[Dict]:
-    """Return jobs ordered by number of positive ratings."""
-    conn = sqlite3.connect(DATABASE)
-    cur = conn.cursor()
-    cur.execute(
-        """
-        SELECT j.*,
-               COALESCE(SUM(CASE WHEN f.liked=1 THEN 1 ELSE 0 END), 0) AS likes,
-               COALESCE(SUM(CASE WHEN f.liked=0 THEN 1 ELSE 0 END), 0) AS dislikes,
-               CASE WHEN s.job_id IS NOT NULL THEN 1 ELSE 0 END AS has_summary,
-               CASE WHEN e.job_id IS NOT NULL THEN 1 ELSE 0 END AS has_embedding
-        FROM jobs j
-        LEFT JOIN feedback f ON j.id = f.job_id
-        LEFT JOIN summaries s ON j.id = s.job_id
-        LEFT JOIN embeddings e ON j.id = e.job_id
-        GROUP BY j.id
-        ORDER BY likes DESC
-        """
-    )
-    rows = cur.fetchall()
-    columns = [c[0] for c in cur.description]
-    conn.close()
-    return [dict(zip(columns, r)) for r in rows]
-
-
-def aggregate_job_stats() -> Dict:
-    """Return aggregate statistics about the stored jobs."""
-    conn = sqlite3.connect(DATABASE)
-    cur = conn.cursor()
-
-    cur.execute("SELECT COUNT(*) FROM jobs")
-    total_jobs = cur.fetchone()[0]
-
-    cur.execute("SELECT site, COUNT(*) FROM jobs GROUP BY site")
-    by_site = {site: count for site, count in cur.fetchall()}
-
-    cur.execute("SELECT date_posted, COUNT(*) FROM jobs GROUP BY date_posted")
-    by_date = {date: count for date, count in cur.fetchall()}
-
-    cur.execute(
-        "SELECT AVG(min_amount), AVG(max_amount) FROM jobs WHERE min_amount IS NOT NULL AND max_amount IS NOT NULL"
-    )
-    avg_min, avg_max = cur.fetchone()
-
-    cur.execute("SELECT COUNT(*) FROM summaries")
-    summaries_count = cur.fetchone()[0]
-    cur.execute("SELECT COUNT(*) FROM embeddings")
-    embeddings_count = cur.fetchone()[0]
-
-    conn.close()
-    return {
-        "total_jobs": total_jobs,
-        "by_site": by_site,
-        "by_date": by_date,
-        "avg_min_pay": avg_min or 0,
-        "avg_max_pay": avg_max or 0,
-        "jobs_with_summaries": summaries_count,
-        "jobs_with_embeddings": embeddings_count,
-    }
-
-
-def increment_rating_count(job_id: int) -> None:
-    """Increment the rating count for a job."""
-    conn = sqlite3.connect(DATABASE)
-    cur = conn.cursor()
-    cur.execute("SELECT rating_count FROM jobs WHERE id=?", (job_id,))
-    row = cur.fetchone()
-    if not row:
-        conn.close()
-        return
-    count = row[0] + 1
-    cur.execute("UPDATE jobs SET rating_count=? WHERE id=?", (count, job_id))
-    conn.commit()
-    conn.close()
-
-
-def record_feedback(job_id: int, liked: bool, reason: Optional[str]) -> None:
-    """Store feedback and increment rating count."""
-    increment_rating_count(job_id)
-    conn = sqlite3.connect(DATABASE)
-    cur = conn.cursor()
-    cur.execute(
-        "INSERT INTO feedback(job_id, liked, reason, rated_at) VALUES(?,?,?,?)",
-        (job_id, int(liked), reason, int(time.time())),
-    )
-    conn.commit()
-    conn.close()
 
 
 def fetch_jobs_task(search_term: str, location: str, sites: List[str]) -> None:
@@ -485,10 +180,12 @@ def fetch_jobs_task(search_term: str, location: str, sites: List[str]) -> None:
 @app.on_event("startup")
 def on_startup() -> None:
     init_db()
+    train_model()
     if OLLAMA_ENABLED:
         def worker():
             ensure_model_downloaded()
             process_all_jobs()
+            train_model()
 
         Thread(target=worker, daemon=True).start()
 
@@ -551,31 +248,13 @@ def feedback(job_id: int = Form(...), liked: int = Form(...), reason: str = Form
 def stats(request: Request):
     jobs = list_jobs_by_feedback()
     agg = aggregate_job_stats()
+    predictions = predict_unrated()
     return templates.TemplateResponse(
-        "stats.html", {"request": request, "jobs": jobs, "stats": agg}
+        "stats.html",
+        {"request": request, "jobs": jobs, "stats": agg, "predictions": predictions},
     )
 
 
-def cleanup_jobs() -> int:
-    """Delete jobs without a description and remove duplicate title/company entries."""
-    conn = sqlite3.connect(DATABASE)
-    cur = conn.cursor()
-
-    cur.execute("DELETE FROM jobs WHERE description IS NULL OR description = ''")
-    deleted_missing = cur.rowcount
-
-    cur.execute(
-        """
-        DELETE FROM jobs
-        WHERE id NOT IN (
-            SELECT MIN(id) FROM jobs GROUP BY title, company
-        )
-        """
-    )
-    deleted_dupes = cur.rowcount
-    conn.commit()
-    conn.close()
-    return deleted_missing + deleted_dupes
 
 
 @app.get("/manage", response_class=HTMLResponse)
