+++ app/templates/base.html
@@ -7,7 +7,7 @@
   <title>{% block title %}Job Ranker{% endblock %}</title>
 </head>
 <body class="d-flex flex-column min-vh-100">
-  <nav class="navbar navbar-expand-lg navbar-light bg-light">
+  <nav class="navbar navbar-expand navbar-light bg-light">
     <div class="container">
       <a class="navbar-brand" href="/">Job Ranker</a>
       <div class="navbar-nav">
